fix gcc warnings

+++ src/token.cpp
@@ -40,7 +40,7 @@ std::shared_ptr<Ex> ListToken::transform() {
   HeadEx exprsCopy(Symbol::PLUS, value.size());
   HeadEx region(Symbol::MULT, value.size());
   std::vector<int> bound(value.size());
-  for (int i = 0; i < value.size() - 1; i++) {
+  for (size_t i = 0; i < value.size() - 1; i++) {
     std::shared_ptr<Token> left = value[i];
     std::shared_ptr<Token> right = value[i + 1];
     if (isStrong(left, right)) {
@@ -53,7 +53,7 @@ std::shared_ptr<Ex> ListToken::transform() {
       bound[i] |= B_END;
     }
   }
-  for (int i = 0; i < value.size(); i++) {
+  for (size_t i = 0; i < value.size(); i++) {
     std::shared_ptr<Token> token = value[i];
     int b = bound[i];
     std::shared_ptr<Ex> transformed = token->transform();
